+++ tests/testthat.R
@@ -1,14 +1,9 @@
 library(testthat)
 library(beautier)
 
-unlink(
-  list.dirs(rappdirs::user_cache_dir(appname = "beautier")),
-  recursive = TRUE
-)
+clear_beautier_cache()
+check_empty_beautier_folder()
 
 test_check("beautier")
 
-testthat::expect_equal(
-  0,
-  length(list.files(rappdirs::user_cache_dir(appname = "beautier")))
-)
+check_empty_beautier_folder()